move from qureg to experimental group
