+++ quest/src/api/qureg.cpp
@@ -26,30 +26,6 @@
 #include <string>
 #include <vector>
 
-
-
-// TODO:
-// move this to experimental
-#ifdef QUEST_COMPILE_ADIOS2
-#include <adios2.h>
-#if QUEST_COMPILE_MPI
-#include <mpi.h>
-#endif
-#endif
-
-#if QUEST_COMPILE_ADIOS2
-// In distributed builds, ADIOS2 must be given QuEST's communicator so that each
-// node's call collectively writes/reads its own slice of the shared file. Without
-// it, ADIOS2 runs serially per rank and the per-node slices never form one file.
-static adios2::ADIOS makeAdios() {
-#if QUEST_COMPILE_MPI
-    return adios2::ADIOS(MPI_COMM_WORLD);
-#else
-    return adios2::ADIOS();
-#endif
-}
-#endif
-
 using std::string;
 using std::vector;
 
@@ -584,116 +560,4 @@ vector<vector<qcomp>> getDensityQuregAmps(Qureg qureg, qindex startRow, qindex s
     // modify out through its ptrs
     getDensityQuregAmps(ptrs.data(), qureg, startRow, startCol, numRows, numCols);
     return out;
-}
-
-
-
-
-
-// TODO:
-// move this to experimental
-
-/*
- * CHECKPOINTING
- *
- * which is compiled only when QUEST_COMPILE_ADIOS2=ON (requiring ADIOS2).
- * The API functions are always defined so that the validation layer can throw
- * a clear error in non-checkpointing builds, rather than failing to link.
- *
- * These are defined with C linkage (matching their extern "C" declarations in
- * qureg.h) so they remain callable from C consumers; the signatures pass no
- * qcomp by value and so stay C-ABI-safe.
- */
-
-
-extern "C" void saveQuregToFile(Qureg qureg, const char* fn) {
-    validate_quregCheckpointingIsCompiled(__func__);
-
-#ifdef QUEST_COMPILE_ADIOS2
-    validate_quregFields(qureg, __func__);
-
-    // ensure the CPU amplitudes reflect any GPU-resident state before writing
-    syncQuregFromGpu(qureg);
-
-    adios2::ADIOS adios = makeAdios();
-    adios2::IO io = adios.DeclareIO("QuESTQuregSave");
-    adios2::Engine engine = io.Open(fn, adios2::Mode::Write);
-
-    // global single-value metadata; we deliberately record only the dimension
-    // and precision, never incidental deployment fields (the loader chooses its
-    // own deployment) nor derivable fields (like numAmps)
-    adios2::Variable<int> vNumQubits  = io.DefineVariable<int>("numQubits");
-    adios2::Variable<int> vIsDensMatr = io.DefineVariable<int>("isDensityMatrix");
-    adios2::Variable<int> vQrealBytes = io.DefineVariable<int>("qrealBytes");
-
-    // amplitudes are stored as interleaved (real, imag) reals to stay agnostic
-    // to precision and to ADIOS2's complex-type support; each node writes only
-    // its local slice into the global array, avoiding excessive memory use
-    qindex globalReals = 2 * qureg.numAmps;
-    qindex localReals  = 2 * qureg.numAmpsPerNode;
-    qindex startReal   = 2 * ((qindex) qureg.rank) * qureg.numAmpsPerNode;
-    adios2::Variable<qreal> vAmps = io.DefineVariable<qreal>(
-        "amps",
-        { (size_t) globalReals },
-        { (size_t) startReal },
-        { (size_t) localReals });
-
-    int qrealBytes = (int) sizeof(qreal);
-
-    engine.BeginStep();
-    engine.Put(vNumQubits,  qureg.numQubits);
-    engine.Put(vIsDensMatr, qureg.isDensityMatrix);
-    engine.Put(vQrealBytes, qrealBytes);
-    engine.Put(vAmps, reinterpret_cast<qreal*>(qureg.cpuAmps));
-    engine.EndStep();
-    engine.Close();
-#endif
-}
-
-
-extern "C" Qureg createQuregFromFile(const char* fn) {
-    validate_quregCheckpointingIsCompiled(__func__);
-
-#ifdef QUEST_COMPILE_ADIOS2
-    adios2::ADIOS adios = makeAdios();
-    adios2::IO io = adios.DeclareIO("QuESTQuregLoad");
-    adios2::Engine engine = io.Open(fn, adios2::Mode::Read);
-
-    engine.BeginStep();
-
-    // read dimension + precision metadata first, so we can size the new Qureg
-    int numQubits = 0;
-    int isDensMatr = 0;
-    int fileQrealBytes = 0;
-    engine.Get(io.InquireVariable<int>("numQubits"),       numQubits);
-    engine.Get(io.InquireVariable<int>("isDensityMatrix"), isDensMatr);
-    engine.Get(io.InquireVariable<int>("qrealBytes"),      fileQrealBytes);
-    engine.PerformGets();
-
-    validate_quregFileMatchesPrecision(fileQrealBytes, __func__);
-
-    // create a matching-dimension Qureg with automatically chosen deployments,
-    // independent of those used when the file was saved
-    Qureg qureg = (isDensMatr)?
-        createDensityQureg(numQubits) :
-        createQureg(numQubits);
-
-    // read only this node's slice of the global amplitude array into its buffer
-    qindex localReals = 2 * qureg.numAmpsPerNode;
-    qindex startReal  = 2 * ((qindex) qureg.rank) * qureg.numAmpsPerNode;
-    adios2::Variable<qreal> vAmps = io.InquireVariable<qreal>("amps");
-    vAmps.SetSelection({ { (size_t) startReal }, { (size_t) localReals } });
-    engine.Get(vAmps, reinterpret_cast<qreal*>(qureg.cpuAmps));
-
-    engine.EndStep();
-    engine.Close();
-
-    // propagate the restored CPU amplitudes to the GPU, if deployed
-    syncQuregToGpu(qureg);
-
-    return qureg;
-#else
-    // unreachable: the validation above always throws in non-checkpointing builds
-    return Qureg{};
-#endif
 }
