+++ tests/unit/CMakeLists.txt
@@ -4,7 +4,6 @@ target_sources(tests
   PUBLIC
   calculations.cpp
   channels.cpp
-  checkpoint.cpp
   debug.cpp
   decoherence.cpp
   environment.cpp
